Refactoring: GeoDataRiverSurvey::MouseMoveExtensionCommand is moved to separated files.
