+++ libs/geodata/riversurvey/riversurvey.pro
@@ -150,6 +150,7 @@ HEADERS += gd_riversurvey_global.h \
            private/geodatariversurvey_expandcrosssectioncommand.h \
            private/geodatariversurvey_insertriverpathpointcommand.h \
            private/geodatariversurvey_mouseeditcrosssectioncommand.h \
+           private/geodatariversurvey_mousemoveextensioncommand.h \
            private/geodatariversurvey_mouserotaterivercrosssectioncommand.h \
            private/geodatariversurvey_mouseshiftriverpathcentercommand.h \
            private/geodatariversurvey_moveriverpathpointcommand.h \
@@ -207,6 +208,7 @@ SOURCES += geodatarivercrosssection.cpp \
            private/geodatariversurvey_expandcrosssectioncommand.cpp \
            private/geodatariversurvey_insertriverpathpointcommand.cpp \
            private/geodatariversurvey_mouseeditcrosssectioncommand.cpp \
+           private/geodatariversurvey_mousemoveextensioncommand.cpp \
            private/geodatariversurvey_mouserotaterivercrosssectioncommand.cpp \
            private/geodatariversurvey_mouseshiftriverpathcentercommand.cpp \
            private/geodatariversurvey_moveriverpathpointcommand.cpp \
